// src/models/quizzr.ts
export class Quizzr{
    
   start_date: any;
   end_date :any;
   quiz_type: any;
   description: any;
   duration_mins: number;
   attempts_allowed :number;
   randomize:boolean;
   title:any;
   question_number:number;
   question: any;
   sequence_number: number;
   answer_type:any;
   correct_answer_count: number;
   explanation: any;
   score: number;
   attempt_id: number;
   answer: any;
   user_id:number;
   is_right_answer:boolean;


}